Move to ui/ directory to reduce clutter in the toplevel directory. Update
* regexxer.{desktop,schemas}.in: Move to ui/ directory to reduce
clutter in the toplevel directory.
* Makefile.am ({desktop,schemas}_in_files): Update accordingly.
({desktop,schemas}_DATA): Likewise.
(ui/stockimages.h): Do not rely on $(mkdir_p) because it is already
obsolete but AC_PROG_MKDIR_P() is too recent to require.  Just use
plain mkdir and test for the existence of the directory on failure.
Also, add a comment explaining the implicit directory creation.

* POTFILES.in: Move regexxer.{desktop,schemas}.in to the ui/
directory.

svn path=/trunk/; revision=582