:wstring_to_utf8): New function to convert an std::wstring to UTF-8. This
* src/stringutils.{cc,h} (Util::wstring_to_utf8): New function to
convert an std::wstring to UTF-8.  This is a temporary measure
to deal with the same locale issues as discussed in the comments
to bug #399216.
(Util::int_to_string): Use a wide character stream and convert
the result to UTF-8 with Util::wstring_to_utf8().
* src/statusline.cc (CounterBox::stringstream_): Change to
std::wostringstream, as above.
(CounterBox::number_to_string): Use Util::wstring_to_utf8().

svn path=/trunk/; revision=650